+++ bin/node-template/runtime/Cargo.toml
@@ -15,6 +15,7 @@ codec = { package = "parity-scale-codec", version = "2.0.0", default-features = 
 
 pallet-aura = { version = "3.0.0", default-features = false, path = "../../../frame/aura" }
 pallet-balances = { version = "3.0.0", default-features = false, path = "../../../frame/balances" }
+frame-accounts = { version = "3.0.0", default-features = false, path = "../../../frame/accounts" }
 frame-support = { version = "3.0.0", default-features = false, path = "../../../frame/support" }
 pallet-grandpa = { version = "3.0.0", default-features = false, path = "../../../frame/grandpa" }
 pallet-randomness-collective-flip = { version = "3.0.0", default-features = false, path = "../../../frame/randomness-collective-flip" }
